deep-researcher-agent/backend/requirements.txt
flask
flask-cors
sentence-transformers
transformers
torch
faiss-cpu
